refactor: move filewidget functions to separate file
