+++ app/lib/theme/widgets/FileWidget.tsx
@@ -2,7 +2,11 @@ import React, { useState } from 'react';
 import { useRouter } from 'next/router';
 import * as Sentry from '@sentry/nextjs';
 import { WidgetProps } from '@rjsf/core';
-import path from 'path';
+import {
+  deleteFileFromFormData,
+  validateFile,
+  handleDownload,
+} from 'lib/theme/functions/fileWidgetFunctions';
 import { useCreateAttachment } from '../../../schema/mutations/attachment/createAttachment';
 import { useDeleteAttachment } from '../../../schema/mutations/attachment/deleteAttachment';
 
@@ -20,13 +24,6 @@ type File = {
 interface FileWidgetProps extends WidgetProps {
   value: Array<File>;
 }
-
-const checkFileType = (file, fileTypes) => {
-  const extension = path.extname(file)?.toLowerCase();
-  const typesArr = fileTypes && fileTypes.replace(/ /g, '').split(',');
-
-  return typesArr.includes(extension);
-};
 
 const FileWidget: React.FC<FileWidgetProps> = ({
   id,
@@ -64,45 +61,20 @@ const FileWidget: React.FC<FileWidgetProps> = ({
       },
     };
 
-    const deleteFileFromFormData = (res) => {
-      const attachmentRowId = res?.updateAttachmentByRowId?.attachment?.rowId;
-      const indexOfFile = value.findIndex(
-        (object) => object.id === attachmentRowId
-      );
-      const newFileList = [...value];
-      newFileList.splice(indexOfFile, 1);
-      const isFileListEmpty = newFileList.length <= 0;
-      onChange(isFileListEmpty ? null : newFileList);
-    };
-
     deleteAttachment({
       variables,
       onError: (res) => {
         /// Allow files to be deleted from form data if attachment record was already archived
         if (res.message.includes('Deleted records cannot be modified')) {
-          deleteFileFromFormData(res);
+          deleteFileFromFormData(res, value, onChange);
         } else {
           setError('deleteFailed');
         }
       },
       onCompleted: (res) => {
-        deleteFileFromFormData(res);
+        deleteFileFromFormData(res, value, onChange);
       },
     });
-  };
-
-  const validateFile = (file: globalThis.File) => {
-    if (!file) return { isValid: false, error: '' };
-
-    const { size } = file;
-    if (size > maxFileSizeInBytes) {
-      return { isValid: false, error: 'fileSize' };
-    }
-    if (acceptedFileTypes && !checkFileType(file.name, acceptedFileTypes)) {
-      return { isValid: false, error: 'fileType' };
-    }
-
-    return { isValid: true, error: null };
   };
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -119,7 +91,11 @@ const FileWidget: React.FC<FileWidgetProps> = ({
       parseInt(router?.query?.applicationId as string, 10);
     const file = e.target.files?.[0];
 
-    const { isValid, error: newError } = validateFile(file);
+    const { isValid, error: newError } = validateFile(
+      file,
+      maxFileSizeInBytes,
+      acceptedFileTypes
+    );
     if (!isValid) {
       setError(newError);
       return;
@@ -179,23 +155,6 @@ const FileWidget: React.FC<FileWidgetProps> = ({
     });
 
     e.target.value = '';
-  };
-
-  const showModal = () => {
-    window.location.hash = 'file-error';
-  };
-
-  const handleDownload = async (uuid, fileName) => {
-    const url = `/api/s3/download/${uuid}/${fileName}`;
-    await fetch(url)
-      .then((response) => response.json())
-      .then((response) => {
-        if (response.avstatus) {
-          showModal();
-        } else {
-          window.open(response, '_blank');
-        }
-      });
   };
 
   return (
